Multiple changes to the notebook for binary contingency tables (#350)
* Multiple changes to the notebook for binary contingency tables

Modify data to be around 50 (mm)
Adjust dimension from height to lead hour
Improve the clarity on the threshold being used. The default operator for the class is now greater-than-or-equal-to. The notebook has been adjusted to make the operator usage clearer.
The title level has been adjusted to level 1 to improve documentation rendering
Add link to further reading